rename trajectory.py to newportxps.py
